+++ src/i3dm.tooling/Commands/ICommand.cs
@@ -1,0 +1,7 @@
+namespace i3dm.tooling.Commands
+{
+    public interface ICommand
+    {
+        void Run();
+    }
+}
